+++ sso/src/schemas/auth/registerSchema.ts
@@ -1,0 +1,20 @@
+import z from 'zod'
+import { userSchema } from './userSchema'
+import { passwordRegex } from '../passwordRegex'
+
+export const registerSchema = userSchema
+  .pick({
+    dni: true,
+    email: true,
+  })
+  .extend({
+    // accept: z.literal<boolean>(true),
+    password: z.string().min(8).regex(passwordRegex),
+    confirmPassword: z.string().min(8).regex(passwordRegex),
+  })
+  .refine((data) => data.password === data.confirmPassword, {
+    message: 'Passwords must match',
+    path: ['confirmPassword'],
+  })
+
+export type UserRegister = z.infer<typeof registerSchema>