+++ sso/src/schemas/index.ts
@@ -1,0 +1,4 @@
+export { loginSchema } from './auth/loginSchema'
+export { registerSchema } from './auth/registerSchema'
+export { userSchema } from './auth/userSchema'
+export { dniSchema } from './dniSchema'