+++ sso/CHANGELOG.md
@@ -1,6 +1,30 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [1.2.0] - 2023-11-22
+
+### Changed
+
+- Updated "users" table to "user"
+
+### Testing
+
+- Added tests for POST validate token.
+
+### Documentation
+
+- Swagger endpoint for POST validate token.
+- Updated description for POST login user.
+  
+## [1.1.1] - 2023-11-20
+
+### Added
+
+- Swagger endpoint for POST register user.
+
+### Fixed
+
+- Register endpoint schema, route, controller and test.
 
 ## [1.1.0] - 2023-11-14